requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.6
Django==4.1.5
django-model-utils==4.3.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
packaging==23.0
psycopg2==2.9.5
python-decouple==3.6
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
